requirements.txt
#django==1.3.1
django==1.4.5
django-lfs==0.7.8
Pillow==2.4.0
South==0.8.4
argparse==1.2.1
dj-database-url==0.2.2
-e git://github.com/kennethreitz/dj-static@e79582483a976c76511244b3f236f7c68991e1ae#egg=dj_static-dev
django-appconf==0.6
#-e git://github.com/MagicSolutions/django-carousel.git#egg=django-carousel
django-countries==1.5
django-paintstore>=0.2
django-hvad==0.4.0
django-image-cropping==0.8.2
easy-thumbnails==2.0.1
gunicorn==18.0
#psycopg2==2.5.1
six==1.6.1
static==0.4
static3==0.5
wsgiref==0.1.2